Had to combine two files to get rid of the cycle dep warning as they are deeply coupled.
